[Data] Organize tests in test_formats.py
Reorganize tests from test_formats.py into more specific test files:
- Move arrow-related tests to new test_arrow.py
- Move iteration tests to test_dataset_iter.py
- Move filesystem/parquet test to test_parquet.py
- Move HTTP filesystem test to test_text.py
- Move example data test to test_csv.py
- Move error handling test to test_file_based_datasource.py

This improves test organization by grouping related tests together.

Related issue: #60349
