+++ python/ray/data/tests/datasource/test_csv.py
@@ -344,6 +344,20 @@ def test_write_min_rows_per_file(
             assert num_rows_written == min_rows_per_file
 
 
+def test_read_example_data(ray_start_regular_shared, tmp_path):
+    ds = ray.data.read_csv("example://iris.csv")
+    assert ds.count() == 150
+    assert ds.take(1) == [
+        {
+            "sepal.length": 5.1,
+            "sepal.width": 3.5,
+            "petal.length": 1.4,
+            "petal.width": 0.2,
+            "variety": "Setosa",
+        }
+    ]
+
+
 if __name__ == "__main__":
     import sys
